+++ src/main/java/net/commoble/exmachina/api/Receiver.java
@@ -5,7 +5,6 @@ import net.minecraft.world.level.LevelAccessor;
 /**
  * Function which is invoked for {@link SignalReceiver}s in a signal graph after a signal graph update occurs.
  */
-@FunctionalInterface
 public interface Receiver
 {
 	/**
@@ -14,4 +13,9 @@ public interface Receiver
 	 * @param power The new signal power level assigned to the receiver node in the range [0,15]
 	 */
 	public abstract void accept(LevelAccessor level, int power);
+	
+	/**
+	 * {@return NodeShape of this receiver}
+	 */
+	public abstract NodeShape nodeShape();
 }