+++ src/main/java/net/commoble/exmachina/api/TransmissionNode.java
@@ -9,15 +9,17 @@ import net.minecraft.world.level.LevelAccessor;
 
 /**
  * Definition of a {@link SignalTransmitter} graph node.
+ * @param shape NodeShape of the node
  * @param powerReaders Set of Directions to read adjacent vanilla redstone signal power into this graph from (only blocks outside of the graph are read from)
- * @param connectableNodes Set of Faces (BlockPos+Direction) which this graph node can connect to
+ * @param connectableNodes Set of Nodes which this graph node can connect to. This should specify the smallest connectable NodeShape; larger nodeshapes at the target pos will form connections, while smaller nodeshapes will not.
  * @param graphListener BiFunction to run when a graph update occurs.
  * The provided integer is the graph's new signal power level in the range [0,15],
  * while the Map indicates what adjacent directions should receive neighbor updates after the signal graph updates.
  */
 public record TransmissionNode(
+	NodeShape shape,
 	Set<Direction> powerReaders,
-	Set<Face> connectableNodes,
+	Set<Node> connectableNodes,
 	BiFunction<LevelAccessor,Integer,Map<Direction, SignalStrength>> graphListener)
 {
 }