+++ src/main/java/net/commoble/exmachina/api/content/DefaultSource.java
@@ -9,13 +9,15 @@ import com.mojang.serialization.MapCodec;
 
 import net.commoble.exmachina.api.Channel;
 import net.commoble.exmachina.api.ExMachinaRegistries;
-import net.commoble.exmachina.api.Face;
+import net.commoble.exmachina.api.Node;
+import net.commoble.exmachina.api.NodeShape;
 import net.commoble.exmachina.api.SignalSource;
 import net.commoble.exmachina.internal.ExMachina;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.world.level.BlockGetter;
+import net.minecraft.world.level.Level;
 import net.minecraft.world.level.LevelReader;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.state.BlockState;
@@ -105,24 +107,40 @@ public enum DefaultSource implements SignalSource
 	}
 
 	@Override
-	public Map<Channel, ToIntFunction<LevelReader>> getSupplierEndpoints(BlockGetter level, BlockPos supplierPos, BlockState supplierState, Direction supplierSide, Face connectedFace)
+	public Map<Channel, ToIntFunction<LevelReader>> getSupplierEndpoints(ResourceKey<Level> supplierLevelKey, BlockGetter supplierLevel, BlockPos supplierPos,
+		BlockState supplierState, NodeShape preferredSupplierShape, Node connectingNode)
 	{
 		// we allow wires to connect to vanilla power emitters by default if the block is redstone-connectable and has a connectable voxelshape
-		BlockPos wirePos = connectedFace.pos();
+		BlockPos wirePos = connectingNode.pos();
 		BlockPos offsetFromNeighbor = supplierPos.subtract(wirePos);
 		@Nullable Direction directionFromNeighbor = Direction.getNearest(offsetFromNeighbor, null); 
-		if (!supplierState.canRedstoneConnectTo(level, wirePos, directionFromNeighbor))
+		if (!supplierState.canRedstoneConnectTo(supplierLevel, wirePos, directionFromNeighbor))
 			return Map.of();
 		Direction directionToWire = directionFromNeighbor.getOpposite();
-		VoxelShape wireTestShape = SMALL_NODE_SHAPES[connectedFace.attachmentSide().ordinal()];
-		VoxelShape neighborShape = supplierState.getBlockSupportShape(level, supplierPos);
-		VoxelShape projectedNeighborShape = neighborShape.getFaceShape(directionToWire);
+		boolean canNodeConnectToSupplier = switch(connectingNode.shape())
+		{
+			case NodeShape.Cube cube -> true;
+			case NodeShape.Side side -> {
+				VoxelShape wireTestShape = SMALL_NODE_SHAPES[side.directionToNeighbor().ordinal()];
+				VoxelShape neighborShape = supplierState.getBlockSupportShape(supplierLevel, supplierPos);
+				VoxelShape projectedNeighborShape = neighborShape.getFaceShape(directionToWire);
+				yield Shapes.joinIsNotEmpty(projectedNeighborShape, wireTestShape, BooleanOp.ONLY_SECOND);
+			}
+			case NodeShape.SideSide sideSide -> {
+				if (sideSide.secondaryDirection() != directionFromNeighbor)
+					yield false;
+				VoxelShape wireTestShape = SMALL_NODE_SHAPES[sideSide.directionToNeighbor().ordinal()];
+				VoxelShape neighborShape = supplierState.getBlockSupportShape(supplierLevel, supplierPos);
+				VoxelShape projectedNeighborShape = neighborShape.getFaceShape(directionToWire);
+				yield Shapes.joinIsNotEmpty(projectedNeighborShape, wireTestShape, BooleanOp.ONLY_SECOND);
+			}
+		};
 		// if the projected neighbor shape entirely overlaps the line shape,
 		// then the neighbor shape can be connected to by the wire
 		// we can test this by doing an ONLY_SECOND comparison on the shapes
 		// if this returns true, then there are places where the second shape is not overlapped by the first
 		// so if this returns false, then we can proceed
-		return Shapes.joinIsNotEmpty(projectedNeighborShape, wireTestShape, BooleanOp.ONLY_SECOND)
+		return canNodeConnectToSupplier
 			? Map.of()
 			: Map.of(Channel.redstone(), reader -> reader.getSignal(supplierPos, directionFromNeighbor));
 	}