+++ src/main/java/net/commoble/exmachina/internal/signal/ReceiverPos.java
@@ -1,5 +1,6 @@
 package net.commoble.exmachina.internal.signal;
 
+import net.commoble.exmachina.api.Channel;
 import net.commoble.exmachina.api.SignalReceiver;
 import net.minecraft.core.BlockPos;
 import net.minecraft.resources.ResourceKey;
@@ -9,9 +10,10 @@ import net.minecraft.world.level.Level;
  * SignalReceiver at a given blockpos
  * @param levelKey Level key where the SignalReceiver is
  * @param pos BlockPos where the SignalReceiver is
+ * @param channel Channel the receiver is receiving on
  * @param receiver SignalReceiver at that position
  */
-public record PosReceiver(ResourceKey<Level> levelKey, BlockPos pos, SignalReceiver receiver)
+public record ReceiverPos(ResourceKey<Level> levelKey, BlockPos pos, Channel channel, SignalReceiver receiver)
 {
 
 }