+++ src/main/java/net/commoble/exmachina/internal/signal/SignalGraph.java
@@ -15,7 +15,8 @@ import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.Nullable;
 
 import net.commoble.exmachina.api.Channel;
-import net.commoble.exmachina.api.Face;
+import net.commoble.exmachina.api.Node;
+import net.commoble.exmachina.api.NodeShape;
 import net.commoble.exmachina.api.Receiver;
 import net.commoble.exmachina.api.SignalStrength;
 import net.commoble.exmachina.api.SignalTransmitter;
@@ -45,32 +46,32 @@ import net.minecraft.world.level.block.state.BlockState;
  * @param power The highest level of redstone power provided to this graph
  */
 @ApiStatus.Internal
-public record SignalGraph(Map<NodePos, TransmissionNode> nodesInGraph, Map<ServerLevel, Set<BlockPos>> blocksInGraph, Map<ServerLevel, List<Receiver>> receiverNodes, int power)
+public record SignalGraph(Map<NodeOnChannel, TransmissionNode> nodesInGraph, Map<ServerLevel, Set<BlockPos>> blocksInGraph, Map<ServerLevel, List<Receiver>> receiverNodes, int power)
 {
 	/**
 	 * Creates a SignalGraph at a given origin context
 	 * @param level ServerLevel where the signal graph is being created
-	 * @param originNodePos NodePos (blockpos + direction + channel) of the origin node
+	 * @param originNodeOnChannel NodePos (blockpos + direction + channel) of the origin node
 	 * @param originNode TransmissionNode at the origin nodepos
 	 * @param knownWirers StateWirer lookup cache
 	 * @param unusedReceivers Receiver listeners which were at requested graph update positions but have not been included in any signal graph yet during this graph update
 	 * @return SignalGraph created from the origin context
 	 */
 	@ApiStatus.Internal
-	public static SignalGraph fromOriginNode(ServerLevel level, NodePos originNodePos, TransmissionNode originNode, Map<ServerLevel, Map<BlockPos, StateWirer>> knownWirers, Map<PosReceiver, Set<Receiver>> unusedReceivers)
+	public static SignalGraph fromOriginNode(ServerLevel level, NodeOnChannel originNodeOnChannel, TransmissionNode originNode, Map<ServerLevel, Map<BlockPos, StateWirer>> knownWirers, Map<ReceiverPos, Map<Receiver, Collection<Node>>> unusedReceivers)
 	{
 		// build graph
 		MinecraftServer server = level.getServer();
-		Map<NodePos, TransmissionNode> nodesInGraph = new HashMap<>();
+		Map<NodeOnChannel, TransmissionNode> nodesInGraph = new HashMap<>();
 		Map<ServerLevel, Set<BlockPos>> blocksInGraph = new HashMap<>();
 		Queue<NodeAtPos> uncheckedNodesInGraph = new LinkedList<>();
 		Map<ServerLevel, List<Receiver>> receiverNodes = new HashMap<>();
 		Function<ServerLevel, Function<BlockPos, StateWirer>> wirerLookup = targetLevel -> targetPos -> knownWirers
 			.computeIfAbsent(targetLevel, key -> new HashMap<>())
 			.computeIfAbsent(targetPos, pos -> StateWirer.getOrDefault(targetLevel, pos));
-		nodesInGraph.put(originNodePos, originNode);
-		blocksInGraph.computeIfAbsent(level, l -> new HashSet<>()).add(originNodePos.face().pos());
-		uncheckedNodesInGraph.add(new NodeAtPos(originNodePos, originNode));
+		nodesInGraph.put(originNodeOnChannel, originNode);
+		blocksInGraph.computeIfAbsent(level, l -> new HashSet<>()).add(originNodeOnChannel.node().pos());
+		uncheckedNodesInGraph.add(new NodeAtPos(originNodeOnChannel, originNode));
 		int maxSize = ExMachina.COMMON_CONFIG.maxSignalGraphSize().getAsInt();
 		
 		int highestPowerFound = 0;
@@ -78,74 +79,86 @@ public record SignalGraph(Map<NodePos, TransmissionNode> nodesInGraph, Map<Serve
 		// iterate over the list of nodes we have added to the graph
 		// "next" node: in the graph but not processed neighbors yet
 		whileLoop:
-		while (uncheckedNodesInGraph.poll() instanceof NodeAtPos(NodePos(Face nextFace, Channel nextChannel), TransmissionNode nextNode))
+		while (uncheckedNodesInGraph.poll() instanceof NodeAtPos(NodeOnChannel(Node nextFace, Channel nextChannel), TransmissionNode nextTransmissionNode))
 		{
 			// process next node's neighbors
 			// "target" node: a node which next node can conceivably connect to, and may form a mutual connection, in which case it should be in the graph
 			for (Channel targetChannel : nextChannel.getConnectableChannels())
 			{
 				// look for target nodes
-				for (Face targetFace : nextNode.connectableNodes())
+				for (Node preferredTargetNode : nextTransmissionNode.connectableNodes())
 				{
-					ResourceKey<Level> targetLevelKey = targetFace.levelKey();
+					ResourceKey<Level> targetLevelKey = preferredTargetNode.levelKey();
 					ServerLevel targetLevel = server.getLevel(targetLevelKey);
 					if (targetLevel == null)
 					{
 						continue;
 					}
 					var localWirerLookup = wirerLookup.apply(targetLevel);
-					NodePos targetNodePos = new NodePos(targetFace, targetChannel);
-					// skip target if it's already in the graph
-					if (!nodesInGraph.containsKey(targetNodePos))
+					// get all the transmission nodes from the target and see if any are compatible
+					BlockPos targetPos = preferredTargetNode.pos();
+					StateWirer targetStateWirer = localWirerLookup.apply(targetPos);
+					BlockState targetState = targetStateWirer.state();
+					SignalTransmitter targetTransmitter = targetStateWirer.transmitter();
+					for (TransmissionNode targetTransmissionNode : targetTransmitter.getTransmissionNodes(targetLevelKey, targetLevel, targetPos, targetState, targetChannel))
 					{
-						BlockPos targetPos = targetFace.pos();
-						StateWirer targetStateWirer = localWirerLookup.apply(targetPos);
-						BlockState targetState = targetStateWirer.state();
-						SignalTransmitter targetTransmitter = targetStateWirer.transmitter();
+						Node targetNode = new Node(targetLevelKey, targetPos, targetTransmissionNode.shape());
+						NodeOnChannel targetNodeOnChannel = new NodeOnChannel(targetNode, targetChannel);
+						// skip target if it's already in the graph
+						if (nodesInGraph.containsKey(targetNodeOnChannel))
+							continue;
 						
-						Direction targetSide = targetFace.attachmentSide();
-						@Nullable TransmissionNode targetNode = targetTransmitter.getTransmissionNodes(targetLevel, targetPos, targetState, targetSide).get(targetChannel);
-						if (targetNode != null)
+						// target node exists and is not currently in the graph; can it connect back?
+						// we can form a connection if A) our preferred node is compatible with the target
+						// and B) the target node provides a preferred node compatible with the nextNode
+						// a preferred node is compatible with a target node if it is included by the target
+						// e.g. a preferred sideside is compatible with a targeted cube, but not vice-versa
+						if (targetNode.isValidFor(preferredTargetNode))
 						{
-							// target node exists! but can it connect back?
-							if (targetNode.connectableNodes().contains(nextFace))
+							for (Node nodePreferredByTarget : targetTransmissionNode.connectableNodes())
 							{
-								// yes
-								nodesInGraph.put(targetNodePos, targetNode);
-								blocksInGraph.computeIfAbsent(targetLevel, l -> new HashSet<>()).add(targetPos);
-								uncheckedNodesInGraph.add(new NodeAtPos(targetNodePos, targetNode));
-								if (nodesInGraph.size() >= maxSize)
-									break whileLoop;
-							}
-						}
-						
-						// check receiver nodes, we need to remember the listeners
-						@Nullable Receiver targetReceiver = targetStateWirer.receiver().getReceiverEndpoint(targetLevel, targetPos, targetState, targetSide, nextFace, targetChannel);
-						if (targetReceiver != null)
-						{
-							receiverNodes.computeIfAbsent(targetLevel, l -> new ArrayList<>()).add(targetReceiver);
-							@Nullable Set<Receiver> receiversOnChannel = unusedReceivers.get(new PosReceiver(targetLevelKey, targetPos, targetStateWirer.receiver()));
-							if (receiversOnChannel != null)
-							{
-								receiversOnChannel.remove(targetReceiver);
-							}
-						}
-						
-						// check supplier nodes too
-						if (highestPowerFound < 15)
-						{
-							powerLoop:
-							for (var channelPower : targetStateWirer.source().getSupplierEndpoints(targetLevel, targetPos, targetState, targetSide, nextFace).entrySet())
-							{
-								if (channelPower.getKey() == targetChannel)
+								if (nextFace.isValidFor(nodePreferredByTarget))
 								{
-									int suppliedPower = channelPower.getValue().applyAsInt(targetLevel);
-									if (suppliedPower > highestPowerFound)
-									{
-										highestPowerFound = suppliedPower;
-										if (highestPowerFound >= 15)
-											break powerLoop;
-									}
+									// yes! we have a match and can connect
+									nodesInGraph.put(targetNodeOnChannel, targetTransmissionNode);
+									blocksInGraph.computeIfAbsent(targetLevel, l -> new HashSet<>()).add(targetPos);
+									uncheckedNodesInGraph.add(new NodeAtPos(targetNodeOnChannel, targetTransmissionNode));
+									if (nodesInGraph.size() >= maxSize)
+										break whileLoop;
+									
+								}
+							}
+						}
+					}
+					
+					// check receiver nodes, we need to remember the listeners
+					NodeShape preferredTargetShape = preferredTargetNode.shape();
+					@Nullable Receiver targetReceiver = targetStateWirer.receiver().getReceiverEndpoint(targetLevelKey, targetLevel, targetPos, targetState, preferredTargetShape, nextFace, targetChannel);
+					if (targetReceiver != null)
+					{
+						receiverNodes.computeIfAbsent(targetLevel, l -> new ArrayList<>()).add(targetReceiver);
+						ReceiverPos receiverPos = new ReceiverPos(targetLevelKey, targetPos, targetChannel, targetStateWirer.receiver());
+						@Nullable Map<Receiver, Collection<Node>> receiversAtPos = unusedReceivers.get(receiverPos);
+						if (receiversAtPos != null)
+						{
+							receiversAtPos.remove(targetReceiver);
+						}
+					}
+					
+					// check supplier nodes too
+					if (highestPowerFound < 15)
+					{
+						powerLoop:
+						for (var channelPower : targetStateWirer.source().getSupplierEndpoints(targetLevelKey, targetLevel, targetPos, targetState, preferredTargetShape, nextFace).entrySet())
+						{
+							if (channelPower.getKey() == targetChannel)
+							{
+								int suppliedPower = channelPower.getValue().applyAsInt(targetLevel);
+								if (suppliedPower > highestPowerFound)
+								{
+									highestPowerFound = suppliedPower;
+									if (highestPowerFound >= 15)
+										break powerLoop;
 								}
 							}
 						}
@@ -175,19 +188,19 @@ public record SignalGraph(Map<NodePos, TransmissionNode> nodesInGraph, Map<Serve
 			powerReaderLoop:
 			for (var entry : nodesInGraph.entrySet())
 			{
-				NodePos nodePos = entry.getKey();
-				TransmissionNode node = entry.getValue();
-				Face nodeFace = nodePos.face();
-				var levelKey = nodeFace.levelKey();
+				NodeOnChannel nodeOnChannel = entry.getKey();
+				TransmissionNode transmissionNode = entry.getValue();
+				Node node = nodeOnChannel.node();
+				var levelKey = node.levelKey();
 				ServerLevel nodeLevel = server.getLevel(levelKey);
 				if (nodeLevel == null)
 				{
 					continue;
 				}
 				SignalGetter signalGetter = new WireIgnoringSignalGetter(nodeLevel, wirerLookup.apply(nodeLevel));
-				BlockPos nodeBlockPos = nodeFace.pos();
+				BlockPos nodeBlockPos = node.pos();
 				Set<BlockPos> blocksInLevel = blocksInGraph.computeIfAbsent(level, l -> new HashSet<>());
-				for (Direction directionToNeighbor : node.powerReaders())
+				for (Direction directionToNeighbor : transmissionNode.powerReaders())
 				{
 					BlockPos neighborPos = nodeBlockPos.relative(directionToNeighbor);
 					if (!blocksInLevel.contains(neighborPos))
@@ -214,7 +227,7 @@ public record SignalGraph(Map<NodePos, TransmissionNode> nodesInGraph, Map<Serve
 	 * @param nodePos NodePos to check whether this graph contains
 	 */
 	@ApiStatus.Internal
-	public boolean hasTransmissionNode(NodePos nodePos)
+	public boolean hasTransmissionNode(NodeOnChannel nodePos)
 	{
 		return nodesInGraph().containsKey(nodePos);
 	}
@@ -244,24 +257,26 @@ public record SignalGraph(Map<NodePos, TransmissionNode> nodesInGraph, Map<Serve
 	 * @return Map of node Faces whose neighbors should be updated, e.g. (pos+west) = the block to the west of pos should receive a neighbor update
 	 */
 	@ApiStatus.Internal
-	public Map<Face, SignalStrength> updateListeners(MinecraftServer server)
+	public Map<ResourceKey<Level>, Map<BlockPos, Map<Direction, SignalStrength>>> updateListeners(MinecraftServer server)
 	{
-		Map<Face, SignalStrength> neighborUpdatingNodes = new HashMap<>();
+		Map<ResourceKey<Level>, Map<BlockPos, Map<Direction, SignalStrength>>> neighborUpdatingNodes = new HashMap<>();
 		
 		for (var entry : this.nodesInGraph.entrySet())
 		{
-			NodePos nodePos = entry.getKey();
-			Face face = nodePos.face();
-			BlockPos nodeBlockPos = face.pos();
-			ResourceKey<Level> levelKey = face.levelKey();
+			NodeOnChannel nodeOnChannel = entry.getKey();
+			Node node = nodeOnChannel.node();
+			BlockPos nodeBlockPos = node.pos();
+			ResourceKey<Level> levelKey = node.levelKey();
 			ServerLevel serverLevel = server.getLevel(levelKey);
 			if (serverLevel == null)
 			{
 				continue;
 			}
-			TransmissionNode node = entry.getValue();
-			node.graphListener().apply(serverLevel, this.power).forEach((directionToNeighbor, signalStrength) -> {
-				neighborUpdatingNodes.merge(new Face(nodeBlockPos, directionToNeighbor, levelKey), signalStrength, SignalStrength::max);
+			TransmissionNode transmissionNode = entry.getValue();
+			transmissionNode.graphListener().apply(serverLevel, this.power).forEach((directionToNeighbor, signalStrength) -> {
+				neighborUpdatingNodes.computeIfAbsent(levelKey, $ -> new HashMap<>())
+					.computeIfAbsent(nodeBlockPos, $ -> new HashMap<>())
+					.merge(directionToNeighbor, signalStrength, SignalStrength::max);
 			});
 		}
 		
@@ -277,7 +292,7 @@ public record SignalGraph(Map<NodePos, TransmissionNode> nodesInGraph, Map<Serve
 		return neighborUpdatingNodes;
 	}
 	
-	private static record NodeAtPos(NodePos pos, TransmissionNode node)
+	private static record NodeAtPos(NodeOnChannel pos, TransmissionNode node)
 	{
 
 	}